[IMP] estate: Chapter 12
Added aility to view all properties of each property type.
The state of the property is shown at the header on the top
Added sorting for all models
Made the property and offer buttons conditional
Added color field to the tag model
Added color decorations for property and offer models based on their
state
Offers and tags can be edited directly from list view
Added a default search for properties and the ability to search for
properties, which have a living area greater than the input.
Added stat button to the property types t o show amount of offers for
each