[IMP] estate: Chapter 10 Property and Offer Actions
Add actions to the properties to change the state to sold or canceled,
raise an UserError if the user try to cancel a sold Property or sold a canceled Property.

Add actions to the offers to change the state to accepted or refused,
if accepted : update the property's buyer and selling price fields.
raise a UserError if the user try to accept an offer for a property that has already a accepted offer