[IMP] estate: Chapter 12-Inline Views Widgets List Order & Stat Button
Add an inline view on the property type form view to display the list of properties related to this type.

Add a status bar widget to the property form view to show and select the state of the property.

Add some model order rules on :
- property (ID desc)
- property offer (Price desc)
- property type (Name)
- property tag (Name)
Add a Manual order rule to the Property Type tree view thanks to a handle widget and a sequence field.

Prevent creation and editing of property types from the property form view
add a property tag color

Hide Sold and Cancel button on the property form view if the property is already sold or canceled.

Hide/Show Garden settings on the property form view depending on the garden field state.

Make the property offer and property tag list views editable

Make the date_availability optional and hidden by default on the property list view

Add some decorations on property and property offer list views (green, bold, muted, red)

Add default filter on the property list view to show only available properties

Add a filter domain to the living_area to show only properties with a living area greater than the researched value.

Add a stat button to the the property type form view to show the offers count and redirect to the offers list view populated with the related offers.