[ADD] installment: created a module and added functionalities
Created a module for managing installments, including a cron job for invoices.
Added a settings section specifically for installments.
Implemented a wizard for adding EMI (Equated Monthly Installments).
Introduced a button to redirect to the document module.
Implemented an invoice schedule that triggers when the delay process exceeds the
 specified days.
Added a smart button for document upload requests.
Developed a wizard for document requests related to sales orders.