[ADD] estate: a new mdoule to manage estate properties
Problem
---------
Users are missing a way to manage real estate properties in odoo.

Objective
---------
- Add a way to create new properties, track their attributes and the
state of the sale (new, offer received, ..., sold).
- Add a way to visualise and filter the properties base on their
attributes and status. And give the user a way to categorize the
properties.
- Add a way to manage offer made by parters on individual properties. Be
able to accept/refuse the offers.
- Create an invoice to the client once a sale is closed.

Solution
---------
Add a new module `estate`. This model is in charge of managing the
properties, tracking their sales state and offers.
Add a new module `estate_account` to create invoices once a sale is
closed.

Closes #162