[IMP] estate: Add a server action for better UX
Problem
---------
Users may want to cancel several houses and flats at once, currently
this is not possible; they have to do it one by one.

Objective
---------
Add a way to cancel several estates in one go.

Solution
---------
Add a server action in the demo data so that it is available for new
users. Old users can create the same server action on their database.

task-123456