[IMP] estate: till chapter 14
Chapter - 12
Learned about how to override create and unlink methods with api.model and api.ondelete decorator.
Learned different types of inheritance in odoo like classical inheritance, prototype inheritance and delegation inheritance.
Added property_ids field to res.user model.
Learned how to use xpath and added property_ids field to the base.view_users_form in a new notebook page.

Chapter - 13
This chapter is about module interaction. We want to create the invoice after property is sold.
To do it we require an invoice application and we need to create a bridge module that adds that functionality.
In the bridge module we have inherited the action_sold method and then we have called create method for invoice in account.move model.

Chapter - 14
This chapter describes a brief of Qweb templates.
Using it we can create dynamic templates. In particular to this chapter we require a template to create a kanban view.
 In the exercise we have grouped by the property type and prevented the drag and drop feature of kanban view.