[ADD] estate: Added computed fiels and onchanges
Problem
---------
Users may want to see the total area after adding the garden area and the living area, as well as the best offer after adding multiple offers
to the property. They also might want to either enter a validity number of days or a deadline, and the other should be computed automatically.

Objective
---------
Add a field (total area) dependant on two fields (garden and living area). Add another field which is the maximum of all offer prices.
Add a relation between the validity date and the deadline fields (after creating them).

Solution
---------
Add a computed field for the total area depending on garden and living area fields. Add a computed field for the best price depending on the
offer prices, computing the maximum of them all. Add an computed field for the date deadline and add both a compute and inverse function for it.

task-xxxxxx