[ADD] estate: Added sold, cancel, accept, and refuse buttons. Alongside their actions.
Problem
---------
Users may want to modify the property status through buttons, and see their actions reflected in the data. Same applies to the offer status.

Objective
---------
Add a sold and cancel button for the property, modify the status based on their activity.
Add accept and refuse buttons for the offers, modify the status based on their activity.
Reflect the actions of offers on the property data.

Solution
---------
Add a sold and cancel button in the header part of the XML file, alongside their functions applied on press.
Add a accept and refuse button in the list view of the offers.
Based on which is clicked, modify the selling price and the buyer of the property as well as the status of the offer.

task-xxxxxx