[ADD] estate: Added SQL and Python constraints
Problem
---------
Users may enter incorrect data like a negative expected or offer price. Also, they may create multiple property types and tags
of the same name. The selling price of a property shouldn't be less than 90% of the expected price.

Objective
---------
Add constraints to the field boxes (expected, selling, and offer price). The property tags and types should have constraints too.
Add constraints to prevent selling prices from being lower than 90% of the expected price.

Solution
---------
Added an SQL constraint checking the expected, selling, and offer price to be positive.
Added an SQL constraint to check the uniqueness of the property tags and types.
Added a Python constraint checking the selling price to be not lower than 90% of the expected price.