[ADD] estate: added new res.user model and estate account model
Added code for new res. user model and added code for the account model
to create an invoice when any estate property is sold. I added code for the
Kanban view for easy movement properties. In the invoice, I added a code for the
commission fee and some extra charges for the selling price.