[ADD] estate:  Added code for security rules and pdf reports
Changed offer code in sub-template to better reusability of code and also
Added new report for use-wise property allocated to them and inherited account
module to make sure the invoice is created or not in the report.
added security rules and defined groups related to all models and made sure
Users have the right access to do CRUD on that.
I also added record rules to define security on specific records and learned about
how to bypass the security