[FIX] warranty:  Fixed sale order line warranty configuration
The fixed code is the warranty button that is connected to the product.product
model, it was creating an issue, and now it is connected to the product.template
created different files for the wizard line for better separation of code and
ease of understanding.
The issue fixed related to the field mismatch of the name issue is fixed now.