[ADD] budget:  Added code for analytic account views
Added code analytic account line and that reference module and did code
for budget review options in the menu now it's able to create duplicate
and archived the old budget.
Added view for analytic view changed the code structure files and moved
code of graphs in the budget line module.