[IMP] estate: Added computed fields in estate module.
Added new field in the property model to compute the best offer ever givn to it. Created a method to compute the total area of the property based on garden area and living area.

Created an onchange method as well for the garden checkbox.

Property offer will now have a validity and deadline date which depend on each other.

Completed chpter no 8 of server framework101.