[IMP] estate: Created buttons for property selling and offer accepting.
Performed the following operations while doing chapter no 9 of server framework 101:

Added 2 buttons for selling or cancelling the property and upon the clicking of these buttons some fields are also updating.

Added 2 buttons in the offers view to accpet or reject an offer and then change some values of property model accordingly.