[ADD] warranty: created a new module to add warranty to an sales order
The new module checks which products in a particular sales order can have
warranty.
A button takes us to the wizard which gives options to select the products that
we want to add warranty.
The warranty line gets deleted if the main product is deleted.
Warranty gets added to the sales order, quotation pdf and the invoice pdf.