[ADD] budget_management: add budget management module
Add budget_management module to manage the budget in the analytic account. Users
can create budgets and manage budget lines inside it. All the budget lines are
related to analytic account and analytic account lines.

The target budget for each account can be set and progress can be seen in the
budget line views. Also, users can revise the budget and a new budget can be
seen in chatter.