[IMP] estate:  Add more configuration to user interface.
Chapter 4: Initialize a new file ir.model.access.csv and provide different
permissions in it.
Chapter 5: In this chapter, we learn about XML files and why Odoo uses them. We
then create the estate_property_views.xml file to define actions. After that, we
create a file called estate_menus.xml to display menus for these actions. We
define the necessary fields and attributes for our real estate interface, then
set some limitations on the fields, such as making them read-only. We also add
default values to some fields and introduce two additional fields: active and
state, providing values for them.
Chapter 6: In this chapter, we understand the different types of views (list,
form, search) in an interface. We first create a list view to display our real
estate properties. Then, we create a form view for editing or updating existing
properties, using some HTML tags for customization. Lastly, we create a search
 view to allow users to filter properties based on different criteria. We also
add filters and groupings to display specific data, such as properties with the
state set to "new" or "offer received". During this process, we learn about the
 use of domains in Odoo.