[IMP] estate:  Added access controls, action, menus and custom views.
In Chapter 4: Created ir.model.access.csv file in data folder given read, write,
create and unlink permissions to the group base.group_user. Allowing users in
that group to perform these actions on the model.

In Chapter 5:Created the estate_property_views.xml file in the  views folder.
Created an action for the model estate.property. Created the three levels of
menus  for the estate.property. Set the selling price as read-only
prevent copying of the availability date and the selling price values.Added the
appropriate default attributes. Added the active field to the estate.property model.
Added a state field to the estate.property model.

In Chapter 6: Defined a list view for the estate.property model in
estate_property_views.xml file. Added a custom form view. Added a custom search
view, Filter and Group By.