[IMP] estate: Added function to compute fields
Implemented compute and inverse function  using @api decorater.
'depends' and 'onchange' decorater is implements here.

Using 'depends' decorater, deadline date is calculated using the validity field
and also inverse function of it is created.
Using 'onchange' decorater, 'garden_area' and 'garden_orientation' is set on
selecting the garden field, and it is set to default on unselecting the garden
field.