[ADD] estate: Created new models and added relation in models.
Created the estate.property.type model, along with an action, menu,
and corresponding view
Introduced the property_type_id, salesperson_id, and buyer_id fields in
estate.property to establish a Many2one relationship and added them to the views.
Created the estate.property.tag model with its menu and action. Also added
tags to the estate.property model.
Created the estate.property.offer model, along with a list view and a form view
displaying the price, partner_id, and status fields.
Added the offer_ids field as a One2many virtual relationship in the
estate.property model and included it in the form view.