[IMP] estate:  Add more features to real estate in  form of models.
Chapter 7: In this chapter, we created different models and established links
between them. First, we created a model called estate_property_type, and then we
 created a variable or XML file for property types, allowing us to define
different types of properties like houses and apartments. In this model, we used
the Many2one concept.

Next, we created a different model named estate_property_tag by following the
same steps as the property type model, with the only difference being the use of
the Many2many concept. With this model, we can define various tags, such as
cozy and renovated.

The final model we created in this chapter is estate_property_offer. This model
follows the same steps, but in this case, we don't need to create an XML file,
 and we used the One2many concept.

Overall, we gained an understanding of the concepts Many2one, Many2many, and
One2many.