[IMP] estate:  Added buttons and its actions (chap - 9)
Implemented Sold and Cancel buttons in the header, with functionality to
update the status based on the current status of the record.

Introduced icon buttons to accept or decline an offer. Upon acceptance, the
selling price and buyer's name are automatically updated based on the accepted
offer. Additionally, ensured that only one offer can be accepted at a time.