[IMP] estate:  Create links between fields in the model.
Chapter 8: In this chapter, we understand the concepts of @api.depends,
compute(), which are basically Python logic. Our first task was to add a new
field in estate.property named total_area, which is the sum of living_area and
garden_area. To achieve this, we used the @api.depends or compute() method.
Our second task was to initialize a new field named best_price in estate_
property, with the goal of getting the highest offer from a buyer. In this, we
used the same steps as in total_area.
Our third task was to understand the inverse function. For that, we created two
fields, validity and date_deadline, in the estate.property.offer model. The
 purpose was to allow either validity or date_deadline to be updated, and when
one is updated, the other gets recalculated. For this, we implemented inverse
 methods.
Our last task was to understand the onchange() method. By implementing this,
when the garden field is set, default values will be given to the garden_area
and garden_orientation fields, or vice versa.
Overall, we learned how to write logic in Python by understanding different
 methods.