[IMP] estate:  Added python constraints (chap - 10)
Added constraints for:
The property expected price must be strictly positive.
The property selling price must be positive.
The offer price must be strictly positive.
The property tag name and property type name must be unique.
Additionally, a constraint has been implemented to ensure that the selling price
is not lower than 90% of the expected price.