[IMP] estate:  Link business logic to action buttons.
Chapter9:Our first task was to add Cancel and Sold buttons to the
estate.property model.The logic behind these buttons ensures that if a property
is cancelled, it cannot be sold, and vice versa. We implemented business logic
methods for the buttons and explored how to raise exceptions.
The second task involved adding Accept and Refuse buttons to the
estate.property.offer model. These buttons were represented with cross and check
icons in the offer form. When clicked, they trigger actions that set the offer
status to accepted or refused by initializing the buttons and corresponding
action methods.
The final task required us to implement business logic such that when an offer
is accepted, the property’s buyer and selling price are set based on the
accepted offer. Moreover, only one offer can be accepted for a property at a
time; all other offers are automatically set to refused.
Overall, we gained an understanding of how to trigger actions through buttons by
writing business logic.