[IMP] estate: implemented offers, computed fields, and onchange logic
The estate.property.offer model has been created to manage offers on properties.
This model has fields for price, status (Accepted/Refused),
partner_id (link to res.partner), and property_id (link to estate.property).
Established a One2many relationship in estate.property so that offers could be
displayed within the property form. Views were created to allow offers to be
managed in list and form views.

In estate.property, implemented computed fields of total_area, which is the sum
of living_area and garden_area, and best_price. Implemented in estate.property.
Offer a computed field of date_deadline which computes the period of validity on
create_date and validity with an inverse function for user-set values. Onchange
to automatically populate default garden_area (10) and orientation (North) only
when garden is selected, and reset them when disabled.