[IMP] estate: implemented actions and constraints for property and offers
In this release, i have embedded action buttons into the real estate module to
facilitate part of the business logic. Now, the estate.property model implements
buttons to set a property as sold or canceled that is, a canceled property
cannot be sold , while a sold property  cannot be canceled.
In the estate.property.offer model, we have also implemented buttons to accept
or refuse an offer. An accepted offer sets the selling price, and the buyer
becomes assigned to the property automatically. To keep transactions
unconflicted, we ensure that only one offer can be accepted per property.

Constraints have been added to several places in the code to enforce data
consistency. SQL constraints impose an expected_price in estate.property that is
strictly positive, a selling_price that is positive, and a price in
estate.property.offer that is strictly positive. Moreover, for
consistency's sake, the name fields in estate.property.tag and
estate.property.type are now unique. On the Python side, we implement
a constraint that requires the selling price not to be below 90% of the expected
selling price by using float_compare() and float_is_zero() for numerical
stability.