[IMP] estate: completed chapter 12 and 13
In chapter 12 learned about the 3 inheritances python ,model and view
inheritance. Learned to override crud functions , learned to inherit model in
which there are two types i.e. traditional and delegation inheritance. We used
traditional inheritance to extend the fields of the parent model. In view
inheritance we inherited the base form view for users to add additional data
like properties associated to that user(salesperson)

In chapter 13 learned about interacting with other modules , we created a link
module to facilitate creation of invoice when a property is sold. We inherited
the estate.property model and extended the set_as_sold method to add
invoice and invoice lines automatically.