[IMP] estate:  create security rules, add demo data, enable chatter in property
Day 6:
Restricted property visibility for Agents to only those belonging to
their company by implementing a multi-company rule.
Updated access control to ensure that only users with the Manager role
can view the settings menu in the estate module.
Added demo data for the property, property_type, and property_offer models
to facilitate testing and development.
Integrated a chatter feature in the property model to track changes in the
property name.