[ADD] estate_account:  Added an invoicing feature (Chap-13)
A new module has been integrated, allowing the Sales Order (SO) of a property to
be directly reflected in the Invoicing app. All relevant data from the
property’s Sales Order will be available, and any additional costs will be
applied to the total amount, which can be viewed within the app.
Additionally, the invoicing module enables users to preview and send the
invoice seamlessly.