[IMP] estate:  Add Data Constraints, UI Enhancements, and Enhance Features.
In Chapter 10, SQL constraints were added to the selling price, expected price,
offer price, property tag, and property type to prevent users from entering
incorrect data. A Python constraint was also implemented to ensure that the
selling price cannot be lower than 90% of the expected price.

In Chapter 11, a One2many field, property_ids, was added to the estate.property.
type model and included in the estate.property.type form view. The status bar
widget was used to display the state of estate.property, and model ordering was
added. A sequence field was introduced in the estate.property.type model, and
the sequence was included in the estate.property.type list view with the
appropriate widget. The property_type_id field was modified to prevent the
creation and editing of a property type from the property form view. A color
field was added to the estate.property.tag model. The invisible attribute was
used to conditionally display header buttons, and the garden area and
orientation fields were made invisible in the estate.property form view when no
garden is present. The Accept and Refuse buttons were made invisible once the
offer state was set. The readonly attribute was applied to prevent adding an
offer when the property state was Offer Accepted, Sold, or Cancelled. The
estate.property.offer and estate.property.tag list views were made editable, and
the date_availability field on the estate.property list view was set as optional
and hidden by default. Decorations were added to both the estate.property and
estate.property.offer list views. The Available filter was set to be selected by
default in the estate.property action. A filter_domain was added to the living
area to include properties with an area equal to or greater than the given
value, and a stat button was introduced to the property type.