[IMP] estate: learned to define module data and implemented chatter
Learned to define module data like master data and demo data
Master data is the data that is installed when initializing the server and it is
imperative for the module to function. It is not possible to opt out from it.
We made estate property type master data and it is added to data field in
the manifest. Meanwhile, demo data is the data which we can opt out from if
necessary and it is there for convenience of developers and sales representatives.
It is added to the demo field in the manifest.