[ADD] estate: learned module data
learned about different types of data, such as master data, which is technical
 data necessary for the module to function properly, and demo data, which
 is used for demonstration purposes. Then, learned about data declaration,
 including how data can be declared in either csv or xml. csv is the easiest
way to declare simple data but has limited features, while xml is used when
 the data structure is more complex. Next, studied data extension, where
using field_c allows adding new fields to an existing model in a new module.
 Also learned about ref, which is used in xml to link records by their xml id,
and eval, which is used in xml to evaluate python expressions inside field
tags. learned about functions, which are needed to execute python code
while loading data. explored X2many Fields, which handle one-to-many
and many-to-many relationships in models