[ADD] estate_account: Create a link module for generating invoices.
In Chapter 13, a new module named estate_account was created, which depends on
both the estate and account modules. An estate_property.py file was added inside
the models folder of the estate_account module, inheriting the estate.property
model. The action_sold method was overridden to return the super() call.
Additionally, an empty account.move record was created within the overridden
method, setting the partner_id from the current estate.property and specifying
the move_type as a Customer Invoice. Invoicing lines were added for each
property sold, applying a 6 percent commission on the selling price along with
an additional administrative fee of 100.00.