[IMP] estate:  Enhance UI with inline views,widgets,ordering and stat buttons.
In Chapter-11:
- Added an inline list view for properties in estate_property_type form view.
- Implemented a status bar widget for estate_property state field(New,
Offer Received, Offer Accepted and Sold) with the help of widget
statusbar_visible attributes in field.
- Defined default ordering for models:
  - estate_property: Descending ID
  - estate_property_offer: Descending Price
  - estate_property_tag: Name
  - estate_property_type: Name
In this we understand the _order attribute takes a string containing a list of
fields which will be used for sorting.
- Added manual ordering with a sequence field in estate_property_type.
- Restricted property_type_id creation/editing in estate_property form and
understand no_create: prevent the creation of related records - hide both the
create and edit dropdown menu items (default: false)
- Enabled color tags for estate.property.tag using a color field.
- Conditionally displayed buttons based on estate_property state with the help
of invisible attribute.
- Hide garden-related fields when no garden exists in estate_property same with
the help of invisible attribute.
- Made offers readonly when property state is Offer_Accepted, Sold, or
Cancelled.
- Enabled inline editing for estate_property_offer and estate_property_tag lists
with editable=bottom.
- Set availability date as optional and hidden by default in estate_property
list view using with optional=hide.
- Added color decorations for property and offer states in list views by
understanding the decoration-danger and decoration-success.
- Applied default Available filter in estate_property action by using
search_default_available with context.
- Modified living area search to filter properties with area ≥ input value
by initializing filter_domain in living_area field.
- Introduced a stat button in estate_property_type to list related offers for
achieving this we understand the concept of related fields.

This improves UI clarity, user experience, and data organization.