[IMP] estate:  Added demo data
Demo data was added using both XML and CSV files. The CSV file was used
to populate demo values for the estate property types, while the XML file was
utilized to add demo data for estate properties and their corresponding offers.
Additionally, the  tag in the XML file was employed to execute a
function that accepts an offer for one of the estate property types. Demo data
for the one2many field was also incorporated using the  attribute.