[IMP] estate:  Added PDF report functionality
The functionality has been implemented to allow downloading PDF reports
for specific properties, including details such as the selling price,
orders associated with the property, and selected offers.
The report is structured using a template, and report actions are employed to
trigger the action button for downloading the PDF. Additionally, sub-templates
have been created to enhance code re-usability. Multiple offers for a specific
salesperson can also be downloaded together.