[IMP] estate: added safeguard your code with unit tests
Test 1 - no one can create an offer for a sold property.
Test 2 - no one can sell a property with no accepted offers on it.
Test 3 - check that selling a property that can be sold is correctly marked as
         sold after selling it.
Test 4 - reset of Garden Area and Orientation when you uncheck the Garden
         checkbox.