[IMP] estate: completed mixins ,pdf report and unit testing
Created chatter by inheriting mail.thread mixin. Posted custom message by
overriding the message_post method. Used the tracking attribute on fields so
they are logged in the chatter whenever they are changed. Created a subtype
using the mail.message.subtype model so that the user can choose to receive
these logs related to this subtype or not.

Created downloadable pdf report using the report actions and qweb templates. The
 templates are used to make how the report will look like and the action is used
to bind the template to a model so that the template can use the fields of that
model and also show as a button to access the report. We also define what the
name of the downloaded file will be in the report action. Also used t-call for
sub-templating, basically if a particular section is being used at a lot of
places then we can make it it's own template and use it in some other template
using t-call. Also inherited template and used xpath to add to an already
existing report.

Wrote tests in the tests directory , created the test class that inherits
TransactionCase so that any changes made in the db are rolled back. Made a
setUpClass method which runs once and creates data which can be used throughout
the class , then created different functions for different testing purposes and
used assertRaises , assertRecordValues etc to make sure the right errors are
being raised if we perform a forbidden action. Also used Form from odoo.tests to
make changes via the form view so that we can test if the onchange functionality
is working property or not.