[IMP] estate: improved the module by modifying the manifest file and removing the views folder
I removed the views folder as long as it's not needed for now, I'll try
to keep the folder structure well-adapted with each part of the tutorials.

I changed some fields in my manifest file because they were not correct, the author field changed
from my name 'Hicham' to 'Odoo S.A.', the version is '1.0' now rather than '1.0.0' as long as
this is a small application and it's much convenient for it to have a short version.

Moreover, I added a license field with the value LGPL-3 which is the Odoo license required
for my module, my module is an open-source one therefore the LGPL-3 is convenient in this case.