[IMP] estate:  Added wizard, controllers, chatter
Added chatter in the form view of property, and made state of the property
trackable. And added wizard for making offer in the multiple property record at
once. Added controller to show properties in the website module, created
template to show properties on the webpage. And added filter in the controller
to filter out properties which has creation date below selected date and added
search option on the property name.