[IMP] estate: completed web controller task
Created a menu "Properties" in website which redirects to "/properties" then
created a controller which accepts "/properties" route and displayed property
details. Added Image field in properties too and displayed Images in the front
-end too. Added a pager so that only 6 properties are visible in one page and
user can click next and previous to navigate through pages. Implemented this
using the limit and offset attributes in search on recordset.  Added a date picker
as well where only the properties created after the selected date are displayed.