[IMP] estate:  Add PDF Reports and Unit Tests
In Chapter Build PDF Reports: Added the property offers report to the Print menu
of the Property views and improved it by including more data. Introduced logic
to ensure that when there are no offers on a property, the table is not created,
and a message is displayed instead. Split the offers table into a separate
template for better organization. Additionally, added a new report for res.users
that allows printing all Real Estate Properties visible in their form view,
including the offers associated with each salesperson's properties. Extended the
property report to include information about the invoice.

In Chapter Safeguard your code with unit tests: Updated the code to prevent
creating an offer for a sold property and selling a property without an accepted
offer. Added unit tests to verify these restrictions. Also included a test to
ensure that a property that can be sold is correctly marked as sold after the
sale. Implemented a test to verify that garden_area and garden_orientation are
reset when the garden checkbox is unticked.